Backend/requirements.txt
aniso8601==8.1.0
APScheduler==3.6.3
beautifulsoup4==4.9.3
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
firebase-admin==4.4.0
Flask==1.1.2
Flask-APScheduler==1.11.0
Flask-Cors==3.0.9
Flask-RESTful==0.3.8
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-cloud-core==1.4.4
google-cloud-firestore==2.0.2
google-cloud-storage==1.33.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.6.2
MarkupSafe==1.1.1
msgpack==1.0.0
proto-plus==1.11.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
requests==2.25.0
rsa==4.6
six==1.15.0
soupsieve==2.0.1
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.2
Werkzeug==1.0.1
